upload single file mvp set up
