+++ src/app/action/upload/saveFile.action.jsx
@@ -1,0 +1,33 @@
+'use server'
+
+import { join } from 'path'
+import { writeFile, mkdir } from 'fs/promises'
+import { existsSync } from 'fs'
+import mime from 'mime'
+
+export async function saveFile(formData) {
+  const file = formData.getAll('files')[0] // getting the first file from files
+  const dir = formData.get('dir')
+  const subDir = formData.get('subDir')
+
+  const bytes = await file.arrayBuffer()
+  const buffer = Buffer.from(bytes)
+  const extension = mime.getExtension(file.type)
+  const path = join(
+    process.cwd(),
+    `/public/${dir}/${subDir}`,
+    `image.${extension}`
+  )
+  const dirExists = existsSync(join(process.cwd(), `/public/${dir}`, subDir))
+
+  if (!dirExists) {
+    mkdir(join(process.cwd(), `/public/${dir}`, subDir), { recursive: true })
+  }
+
+  await writeFile(path, buffer)
+
+  return {
+    status: 201,
+    body: path,
+  }
+}
