+++ src/components/UploadFileModal/Uploader/index.jsx
@@ -10,8 +10,8 @@ import { useState } from 'react'
 import Uppy from '@uppy/core'
 import { useDispatch, useSelector } from 'react-redux'
 import { LANGUAGE } from 'app/utils/languages.util'
-import UppyServerActionUpload from 'app/plugins/uppy/UppyServerActionUpload'
-import { saveFiles } from 'app/action/upload/route'
+import UppyServerActionUpload from 'app/plugins/UploadFile/UppyServerActionUpload'
+import { saveFile } from 'app/action/upload/saveFile.action'
 import { useUpdateUserPhoto } from 'app/hooks/user/useUpdateUserPhoto/useUpdateUserPhoto'
 
 export const UppyUploader = () => {
@@ -19,53 +19,28 @@ export const UppyUploader = () => {
   const { lang } = useSelector((state) => state.systemLanguage)
   const { userTable } = useSelector((state) => state.auth)
   const { updateUserPhoto } = useUpdateUserPhoto()
-  const [fileTypes, setFileTypes] = useState([])
-  const [fileNames, setFileNames] = useState([])
 
-  const [uppy] = useState(() =>
-    new Uppy({
-      autoProceed: false,
-      allowMultipleUploadBatches: false,
-      locale: lang === LANGUAGE.uz ? Uzbek : Russian,
-      restrictions: {
-        maxFileSize: 5242880, //5mb
-        allowedFileTypes: ['image/png', 'image/jpeg', 'image/webp'],
-        maxNumberOfFiles: 1,
-        minNumberOfFiles: 1,
-      },
-    })
-      .on('file-added', (result) => {
-        setFileTypes((prev) => [...prev, result?.data?.type])
-        setFileNames((prev) => [...prev, userTable?.id.toString()])
+  const [uppy] = useState(
+    () =>
+      new Uppy({
+        autoProceed: false,
+        allowMultipleUploadBatches: false,
+        locale: lang === LANGUAGE.uz ? Uzbek : Russian,
+        restrictions: {
+          maxFileSize: 5 * 1024 * 1024, // 5 mb
+          allowedFileTypes: ['image/png', 'image/jpeg'],
+          maxNumberOfFiles: 1,
+          minNumberOfFiles: 1,
+        },
+      }).use(UppyServerActionUpload, {
+        action: saveFile,
+        dir: 'user',
+        subDir: userTable?.id.toString() || '',
       })
-      // .on('upload-success', () => {
-      //   setFileTypes([])
-      //   setFileNames([])
-      // })
-      // .on('upload-error', () => {
-      //   setFileTypes([])
-      //   setFileNames([])
-      // })
-      .use(UppyServerActionUpload, {
-        action: saveFiles,
-        dir: 'locales',
-        folderNames: fileNames || [],
-      })
+    // .on('upload-success', () => {
+    //   setFileTypes([])
+    // })
   )
-  // .on('upload-success', () => {
-  //   updateUserPhoto()
-  // })
 
-  //   onBeforeFileAdded: (currentFile) => {
-  //     const modifiedFile = {
-  //       ...currentFile,
-  //       name: `${userTable?.id}.${currentFile.extension}`,
-  //     }
-  //     modifiedFile.name &&
-  //       dispatch(setUserTempData({ photo: modifiedFile.name }))
-
-  //     return modifiedFile
-  //   },
-  // })
   return <Dashboard className="w-full rounded-xl" theme="dark" uppy={uppy} />
 }
